:art: Radio types
